[mod] enum TYPE을 ResourceHandler 밖으로 분리
